+++ core/src/gba/cpu.cpp
@@ -2,165 +2,12 @@
 #include <fmt/printf.h>
 #include <type_traits>
 #include "error_handling.h"
+#include "gba/common_instructions.h"
 #include "gba/hle.h"
+#include "gba/thumb_instructions.h"
 #include "utils.h"
 
 namespace gb::advance {
-enum class ShiftType : u32 {
-  LogicalLeft = 0,
-  LogicalRight,
-  ArithmeticRight,
-  RotateRight,
-};
-
-struct ShiftResult {
-  ShiftType shift_type;
-  u32 shift_amount;
-  u32 shift_value;
-  std::optional<u32> result;
-  std::optional<bool> set_carry;
-};
-
-constexpr ShiftResult compute_shift_value(u32 value, const Cpu& cpu) {
-  const auto shift_type = static_cast<ShiftType>((value >> 5) & 0b11);
-  const auto reg = static_cast<Register>(value & 0xf);
-  const bool register_specified =
-      gb::test_bit(value, 4);  // Shift is specified by a register
-
-  const u8 shift_amount = [&]() -> u8 {
-    if (register_specified) {
-      // Shift by bottom byte of register
-      const u32 shift_reg = (value >> 8) & 0xf;
-      const u8 amount =
-          shift_reg == 15
-              ? 0
-              : static_cast<u8>(cpu.reg(static_cast<Register>(shift_reg)) &
-                                0xff);
-      return amount;
-    }
-
-    // Shift by 5 bit number
-    return static_cast<u8>((value >> 7) & 0b0001'1111);
-  }();
-
-  const u32 reg_value =
-      cpu.reg(reg) +
-      // TODO: Is this offset correct?
-      (register_specified && reg == Register::R15 ? 4 : 0);  // Rm
-
-  if (shift_amount > 0 && shift_amount < 32) {
-    return {shift_type, shift_amount, reg_value, std::nullopt, std::nullopt};
-  }
-  // Allow carry and the result to be overridden
-  const auto [error_value, set_carry] =
-      [&]() -> std::pair<std::optional<u32>, std::optional<bool>> {
-    if (register_specified && shift_amount == 0) {
-      return {{}, cpu.program_status().carry()};
-    }
-    switch (shift_type) {
-      case ShiftType::LogicalLeft:
-        if (register_specified) {
-          if (shift_amount == 32) {
-            return {0, gb::test_bit(reg_value, 0)};
-          }
-          if (shift_amount > 32) {
-            return {0, false};
-          }
-        } else {
-          if (shift_amount == 0) {
-            // return {0, gb::test_bit(reg_value, 31)};
-            return {reg_value, cpu.program_status().carry()};
-          }
-        }
-        return {std::nullopt, std::nullopt};
-
-      case ShiftType::LogicalRight:
-        if (register_specified) {
-          if (shift_amount == 32) {
-            return {0, gb::test_bit(reg_value, 31)};
-          }
-          if (shift_amount > 32) {
-            return {0, false};
-          }
-        } else {
-          if (shift_amount == 0) {
-            return {0, gb::test_bit(reg_value, 31)};
-            // return {reg_value, gb::test_bit(reg_value, 31)};
-          }
-        }
-        return {std::nullopt, std::nullopt};
-      case ShiftType::ArithmeticRight:
-        if (shift_amount >= 32 || shift_amount == 0) {
-          const bool is_negative = gb::test_bit(reg_value, 31);
-          return {is_negative ? 0xffffffff : 0, is_negative};
-        }
-        return {std::nullopt, std::nullopt};
-
-      case ShiftType::RotateRight:
-        if (shift_amount == 0) {
-          const bool carry = gb::test_bit(reg_value, 0);
-          return {(cpu.carry() << 31) | (reg_value >> 1), carry};
-        }
-        return {std::nullopt, std::nullopt};
-
-      default:
-        return {std::nullopt, std::nullopt};
-    }
-  }();
-
-  return {shift_type, shift_amount, reg_value, error_value, set_carry};
-}
-constexpr bool compute_carry(ShiftType shift_type,
-                             u32 shift_operand,
-                             u32 shift_amount) {
-  switch (shift_type) {
-    case ShiftType::LogicalLeft:
-      return shift_amount != 0 &&
-             gb::test_bit(shift_operand, 32 - shift_amount);
-    case ShiftType::LogicalRight:
-      return gb::test_bit(shift_operand, shift_amount - 1);
-    case ShiftType::ArithmeticRight:
-      return gb::test_bit(shift_operand, shift_amount - 1);
-    case ShiftType::RotateRight:
-      return gb::test_bit(shift_operand, shift_amount - 1);
-  }
-  GB_UNREACHABLE();
-}
-
-constexpr u32 compute_result(ShiftType shift_type,
-                             u32 shift_operand,
-                             u32 shift_amount) {
-  switch (shift_type) {
-    case ShiftType::LogicalLeft:
-      return shift_operand << shift_amount;
-    case ShiftType::LogicalRight:
-      return shift_operand >> shift_amount;
-    case ShiftType::ArithmeticRight:
-      return arithmetic_shift_right(shift_operand, shift_amount);
-    case ShiftType::RotateRight:
-      return rotate_right(shift_operand, shift_amount);
-  }
-  GB_UNREACHABLE();
-}
-
-constexpr std::tuple<bool, u32> compute_shifted_operand(
-    const ShiftResult shift_result) {
-  const auto [shift_type, shift_amount, shift_operand, result, set_carry] =
-      shift_result;
-
-  return {
-      set_carry.value_or(
-          compute_carry(shift_type, shift_operand, shift_amount)),
-      result.value_or(compute_result(shift_type, shift_operand, shift_amount))};
-}
-
-static void intr_wait(Cpu& cpu, u32 set_flags, u32 interrupt_flags) {
-  cpu.ime = 1;
-
-  if (set_flags != 0 || cpu.interrupts_waiting.data() == 0) {
-    cpu.interrupts_waiting.set_data(static_cast<u16>(interrupt_flags));
-  }
-}
 
 enum class SoftwareInterruptType : u32 {
   SoftReset = 0x0,
@@ -182,7 +29,7 @@ enum class SoftwareInterruptType : u32 {
   NonstdPrintInt = 0xfe,
 };
 
-static u32 execute_software_interrupt(Cpu& cpu, u32 instruction) {
+u32 execute_software_interrupt(Cpu& cpu, u32 instruction) {
   using namespace gb::advance::hle::bios;
 
   const auto interrupt_type = static_cast<SoftwareInterruptType>(
@@ -225,7 +72,7 @@ static u32 execute_software_interrupt(Cpu& cpu, u32 instruction) {
     case SoftwareInterruptType::ArcTan2: {
       const s16 x = static_cast<s16>(cpu.reg(Register::R0) & 0xffff);
       const s16 y = static_cast<s16>(cpu.reg(Register::R1) & 0xffff);
-      cpu.set_reg(Register::R0, arctan2(x, y));
+      cpu.set_reg(Register::R0, static_cast<u32>(arctan2(x, y)) & 0xffff);
       break;
     }
     case SoftwareInterruptType::CpuSet:
@@ -251,8 +98,11 @@ static u32 execute_software_interrupt(Cpu& cpu, u32 instruction) {
       obj_affine_set(*cpu.mmu(), src, dest, count, stride);
     } break;
     case SoftwareInterruptType::Lz77Wram: {
-      const auto [source_storage, source_addr] =
-          cpu.mmu()->select_storage(cpu.reg(Register::R0));
+      const u32 src = cpu.reg(Register::R0);
+      if (memory_region(src) == 0) {
+        break;
+      }
+      const auto [source_storage, source_addr] = cpu.mmu()->select_storage(src);
       const auto [dest_storage, dest_addr] =
           cpu.mmu()->select_storage(cpu.reg(Register::R1));
       lz77_decompress(source_storage.subspan(source_addr),
@@ -287,7 +137,8 @@ static u32 execute_software_interrupt(Cpu& cpu, u32 instruction) {
       fmt::printf("%d\n", cpu.reg(Register::R0));
       break;
     default:
-      throw std::runtime_error("unimplemented swi");
+      throw std::runtime_error(fmt::format("unimplemented swi {}",
+                                           static_cast<u32>(interrupt_type)));
   }
 
   // 2S + 1N
@@ -370,286 +221,6 @@ void Cpu::handle_interrupts() {
   throw std::runtime_error("invalid condition");
 }
 
-struct ShiftFlags {
-  std::optional<u32> result;
-  bool set_carry;
-};
-
-ShiftFlags compute_carry(Cpu& cpu,
-                         ShiftType shift_type,
-                         bool register_specified,
-                         u32 reg_value,
-                         u32 shift_amount) {
-  const auto [error_value, set_carry] =
-      [&]() -> std::pair<std::optional<u32>, std::optional<bool>> {
-    if (register_specified && shift_amount == 0) {
-      return {{}, cpu.program_status().carry()};
-    }
-    switch (shift_type) {
-      case ShiftType::LogicalLeft:
-        if (register_specified) {
-          if (shift_amount == 32) {
-            return {0, gb::test_bit(reg_value, 0)};
-          }
-          if (shift_amount > 32) {
-            return {0, false};
-          }
-        } else {
-          if (shift_amount == 0) {
-            return {reg_value, cpu.program_status().carry()};
-          }
-        }
-        return {std::nullopt, std::nullopt};
-
-      case ShiftType::LogicalRight:
-        if (register_specified) {
-          if (shift_amount == 32) {
-            return {0, gb::test_bit(reg_value, 31)};
-          }
-          if (shift_amount > 32) {
-            return {0, false};
-          }
-        } else {
-          if (shift_amount == 0) {
-            return {0, gb::test_bit(reg_value, 31)};
-          }
-        }
-        return {std::nullopt, std::nullopt};
-      case ShiftType::ArithmeticRight:
-        if (shift_amount >= 32 || shift_amount == 0) {
-          const bool is_negative = gb::test_bit(reg_value, 31);
-          return {is_negative ? 0xffffffff : 0, is_negative};
-        }
-        return {std::nullopt, std::nullopt};
-
-      case ShiftType::RotateRight:
-        if (shift_amount == 0) {
-          const bool carry = gb::test_bit(reg_value, 0);
-          return {(cpu.carry() << 31) | (reg_value >> 1), carry};
-        }
-        return {std::nullopt, std::nullopt};
-
-      default:
-        return {std::nullopt, std::nullopt};
-    }
-  }();
-  return {error_value, set_carry.value_or([&]() -> bool {
-            switch (shift_type) {
-              case ShiftType::LogicalLeft:
-                return shift_amount != 0 &&
-                       gb::test_bit(reg_value, 32 - shift_amount);
-              case ShiftType::LogicalRight:
-                return gb::test_bit(reg_value, shift_amount - 1);
-              case ShiftType::ArithmeticRight:
-                return gb::test_bit(reg_value, shift_amount - 1);
-              case ShiftType::RotateRight:
-                return gb::test_bit(reg_value, shift_amount - 1);
-            }
-            GB_UNREACHABLE();
-          }())};
-}
-namespace common {
-template <Opcode opcode, bool SetConditionCode>
-void data_processing(Cpu& cpu,
-                     Register dest_reg,
-                     u32 operand1,
-                     u32 operand2,
-                     bool shift_carry) {
-  const auto write_result = [&cpu](Register dest, u32 result) {
-    if (dest == Register::R15 && SetConditionCode) {
-      cpu.move_spsr_to_cpsr();
-    }
-    cpu.set_reg(dest, result);
-  };
-
-  const auto run_logical = [&, shift_carry](u32 result, bool write = true) {
-    if constexpr (SetConditionCode) {
-      cpu.set_carry(shift_carry);
-      cpu.set_zero(result == 0);
-      cpu.set_negative(gb::test_bit(result, 31));
-    } else {
-      static_cast<void>(shift_carry);
-    }
-    if (write) {
-      write_result(dest_reg, result);
-    }
-  };
-
-  const auto run_arithmetic = [&](u64 op1, u64 op2, bool write,
-                                  FunctionRef<u64(u64, u64)> impl,
-                                  bool invert_carry,
-                                  bool carry_override = false,
-                                  bool carry_override_value = false) {
-    const u64 result = impl(op1, op2);
-
-    const u32 result_32 = static_cast<u32>(result & 0xffffffff);
-    if constexpr (SetConditionCode) {
-      const bool overflow =
-          invert_carry
-              ? (((op1 ^ op2) >> 31) & 0b1) && (((op1 ^ result_32) >> 31) & 0b1)
-              : (!(((op1 ^ op2) >> 31) & 0b1)) &&
-                    (((op2 ^ result_32) >> 31) & 0b1);
-      cpu.set_overflow(overflow);
-      cpu.set_zero(result_32 == 0);
-      cpu.set_negative(gb::test_bit(result_32, 31));
-      if (carry_override) {
-        cpu.set_carry(carry_override_value);
-      } else {
-        cpu.set_carry((invert_carry ? (op2 <= op1) : gb::test_bit(result, 32)));
-      }
-    } else {
-      static_cast<void>(carry_override_value);
-    }
-
-    if (write) {
-      write_result(dest_reg, result_32);
-    }
-  };
-  const u32 carry_value = cpu.carry();
-
-  switch (opcode) {
-    // Arithmetic
-    case Opcode::Sub:
-      run_arithmetic(
-          operand1, operand2, true, [](u64 op1, u64 op2) { return op1 - op2; },
-          true);
-      break;
-    case Opcode::Rsb:
-      // op2 - op1
-      run_arithmetic(
-          operand2, operand1, true, [](u64 op1, u64 op2) { return op1 - op2; },
-          true);
-      break;
-    case Opcode::Add:
-      run_arithmetic(
-          operand1, operand2, true, [](u64 op1, u64 op2) { return op1 + op2; },
-          false);
-      break;
-    case Opcode::Adc:
-      run_arithmetic(
-          operand1, operand2, true,
-          [carry_value](u64 op1, u64 op2) { return op1 + op2 + carry_value; },
-          false);
-      break;
-    case Opcode::Sbc:
-      run_arithmetic(
-          operand1, operand2, true,
-          [carry_value](u64 op1, u64 op2) {
-            return op1 - op2 + carry_value - 1;
-          },
-          true, true,
-          (static_cast<u64>(operand2) - static_cast<u64>(carry_value) + 1) <=
-              operand1);
-      break;
-    case Opcode::Rsc:
-      run_arithmetic(
-          operand1, operand2, true,
-          [carry_value](u64 op1, u64 op2) {
-            return op2 - op1 + carry_value - 1;
-          },
-          true, true,
-          static_cast<u64>(operand1) + static_cast<u64>(carry_value) - 1 <=
-              operand2);
-      break;
-    case Opcode::Cmp:
-      run_arithmetic(
-          operand1, operand2, false, [](u64 op1, u64 op2) { return op1 - op2; },
-          true);
-      break;
-    case Opcode::Cmn:
-      run_arithmetic(
-          operand1, operand2, false, [](u64 op1, u64 op2) { return op1 + op2; },
-          false);
-      break;
-    // Logical
-    case Opcode::And:
-      run_logical(operand1 & operand2);
-      break;
-    case Opcode::Eor:
-      run_logical(operand1 ^ operand2);
-      break;
-    case Opcode::Tst:
-      run_logical(operand1 & operand2, false);
-      break;
-    case Opcode::Teq:
-      run_logical(operand1 ^ operand2, false);
-      break;
-    case Opcode::Orr:
-      run_logical(operand1 | operand2);
-      break;
-    case Opcode::Mov:
-      run_logical(operand2);
-      break;
-    case Opcode::Bic:
-      run_logical(operand1 & ~operand2);
-      break;
-    case Opcode::Mvn:
-      run_logical(~operand2);
-      break;
-  }
-}
-void run_shift(Cpu& cpu,
-               ShiftType shift_type,
-               Register dest_reg,
-               u32 value,
-               u32 shift_amount,
-               bool register_specified) {
-  const u32 shifted_value = [&]() -> u32 {
-    switch (shift_type) {
-      // LSL
-      case ShiftType::LogicalLeft:
-        return shift_amount < 32 ? (value << shift_amount) : 0;
-      case ShiftType::LogicalRight:
-        return value >> shift_amount;
-      case ShiftType::ArithmeticRight:
-        return static_cast<s32>(value) >> shift_amount;
-      case ShiftType::RotateRight:
-        return rotate_right(value, shift_amount);
-      default:
-        GB_UNREACHABLE();
-    }
-  }();
-
-  const auto [override_result, carry] =
-      compute_carry(cpu, shift_type, register_specified, value, shift_amount);
-
-  cpu.set_negative(test_bit(shifted_value, 31));
-  cpu.set_zero(override_result.value_or(shifted_value) == 0);
-  cpu.set_carry(carry);
-
-  cpu.set_reg(dest_reg, override_result.value_or(shifted_value));
-}
-
-void multiply(Cpu& cpu,
-              Register dest_register,
-              Register lhs_register,
-              Register rhs_register,
-              bool set_condition_code,
-              std::optional<Register> accumulate_register = std::nullopt) {
-  const u32 rhs_operand = cpu.reg(rhs_register);
-  const u32 res = static_cast<u32>(
-      static_cast<u64>(cpu.reg(lhs_register)) * static_cast<u64>(rhs_operand) +
-      (accumulate_register ? cpu.reg(*accumulate_register) : 0));
-
-  cpu.set_reg(dest_register, res);
-
-  if (set_condition_code) {
-    cpu.set_zero(res == 0);
-    cpu.set_negative(gb::test_bit(res, 31));
-  }
-}
-
-void branch_and_exchange(Cpu& cpu, Register next_pc_reg) {
-  const u32 reg_value = cpu.reg(next_pc_reg);
-
-  const bool thumb_mode = gb::test_bit(reg_value, 0);
-  cpu.set_thumb(thumb_mode);
-
-  cpu.set_reg(Register::R15, reg_value);
-}
-
-}  // namespace common
-
 using InstFunc = u32 (*)(Cpu&, u32);
 
 template <bool link>
@@ -704,93 +275,6 @@ auto compute_operand2(const Cpu& cpu, u32 instruction)
       shift_value<immediate_operand>(cpu, instruction);
   return compute_shifted_operand(shift_result);
 }
-
-static constexpr Cycles load_store_cycles(Register dest_reg, bool load) {
-  if (!load) {
-    return 2_nonseq;
-  }
-
-  // Load
-  return dest_reg == Register::R15 ? (2_seq + 2_nonseq + 1_intern)
-                                   : (1_seq + 1_nonseq + 1_intern);
-}
-
-template <typename T>
-constexpr void run_load(Cpu& cpu, u32 addr, Register dest_reg) {
-  constexpr auto mask = sizeof(T) - 1;
-  const auto aligned_addr = addr & ~mask;
-  const auto rotate_amount = (addr & mask) * 8;
-
-  const auto value = [&] {
-    if constexpr (std::is_same_v<T, s16>) {
-      if ((addr & mask) != 0) {
-        return static_cast<s16>(cpu.mmu()->at<s8>(addr));
-      }
-    }
-    return cpu.mmu()->at<T>(aligned_addr);
-  }();
-  cpu.set_reg(dest_reg,
-              static_cast<std::conditional_t<std::is_signed_v<T>, s32, u32>>(
-                  rotate_right(value, rotate_amount)));
-}
-
-namespace multiply::detail {
-
-[[nodiscard]] constexpr Cycles multiply_cycles(u32 rhs_operand,
-                                               bool accumulate) {
-  const u32 multiply_cycles = [&]() -> u32 {
-    if (const u32 masked = rhs_operand & 0xffffff00;
-        masked == 0 || masked == 0xffffff00) {
-      return 1;
-    }
-    if (const u32 masked = rhs_operand & 0xffff0000;
-        masked == 0 || masked == 0xffff0000) {
-      return 2;
-    }
-
-    if (const u32 masked = rhs_operand & 0xff000000;
-        masked == 0 || masked == 0xff000000) {
-      return 3;
-    }
-
-    return 4;
-  }();
-
-  return Cycles{1, 0, multiply_cycles + (accumulate ? 1 : 0)};
-}
-[[nodiscard]] constexpr Cycles multiply_long_cycles(u32 rhs_operand,
-                                                    bool accumulate,
-                                                    bool is_signed) {
-  if (is_signed) {
-    return multiply_cycles(rhs_operand, accumulate);
-  }
-
-  const u32 multiply_cycles = [rhs_operand] {
-    if (const u32 masked = rhs_operand & 0xffffff00; masked == 0) {
-      return 1;
-    }
-
-    if (const u32 masked = rhs_operand & 0xffff0000; masked == 0) {
-      return 2;
-    }
-
-    if (const u32 masked = rhs_operand & 0xff000000; masked == 0) {
-      return 3;
-    }
-
-    return 4;
-  }();
-
-  return {1, 0, multiply_cycles + (accumulate ? 2 : 1)};
-}
-template <typename T>
-[[nodiscard]] constexpr T multiply(T lhs, T rhs) {
-  static_assert(std::is_integral_v<T>);
-
-  return lhs * rhs;
-}
-
-}  // namespace multiply::detail
 
 namespace arm {
 template <bool accumulate, bool set_condition_code>
@@ -1252,552 +736,7 @@ u32 software_interrupt(Cpu& cpu, u32 instruction) {
 }
 }  // namespace arm
 
-namespace thumb {
-template <ShiftType shift_type, u32 shift_amount>
-u32 move_shifted_register(Cpu& cpu, u16 instruction) {
-  const auto dest_reg = static_cast<Register>(instruction & 0b111);
-  const auto src_reg = static_cast<Register>((instruction >> 3) & 0b111);
-
-  const u32 value = cpu.reg(src_reg);
-  common::run_shift(cpu, shift_type, dest_reg, value, shift_amount, false);
-
-  return 1;
-}
-
-template <bool immediate_operand, bool subtract, u32 register_or_value>
-u32 add_subtract(Cpu& cpu, u16 instruction) {
-  const auto dest_reg = static_cast<Register>(instruction & 0b111);
-  const auto src_reg = static_cast<Register>((instruction >> 3) & 0b111);
-
-  const u32 value = immediate_operand
-                        ? register_or_value
-                        : cpu.reg(static_cast<Register>(register_or_value));
-
-  common::data_processing<subtract ? Opcode::Sub : Opcode::Add, true>(
-      cpu, dest_reg, cpu.reg(src_reg), value, cpu.program_status().carry());
-  return 1;
-}
-
-template <Opcode opcode, Register dest_reg>
-u32 move_compare_add_subtract_immediate(Cpu& cpu, u16 instruction) {
-  const u32 immediate = instruction & 0xff;
-  common::data_processing<opcode, true>(cpu, dest_reg, cpu.reg(dest_reg),
-                                        immediate,
-                                        cpu.program_status().carry());
-  return 1;
-}
-
-template <u32 opcode>
-u32 alu_operation(Cpu& cpu, u16 instruction) {
-  const auto dest_reg = static_cast<Register>(instruction & 0b111);
-  const auto src_reg = static_cast<Register>((instruction >> 3) & 0b111);
-
-  if constexpr (opcode == 0b1101) {
-    // Mul
-    const auto cycles =
-        multiply::detail::multiply_cycles(cpu.reg(dest_reg), false);
-    common::multiply(cpu, dest_reg, src_reg, dest_reg, true);
-    return cycles.sum();
-
-  } else if constexpr ((opcode >= 0b0010 && opcode <= 0b0100) ||
-                       opcode == 0b0111) {
-    // Shift
-    constexpr ShiftType shift_type = []() -> ShiftType {
-      switch (opcode) {
-        case 0b0010:
-          return ShiftType::LogicalLeft;
-        case 0b0011:
-          return ShiftType::LogicalRight;
-        case 0b0100:
-          return ShiftType::ArithmeticRight;
-        case 0b0111:
-          return ShiftType::RotateRight;
-      }
-    }();
-
-    common::run_shift(cpu, shift_type, dest_reg, cpu.reg(dest_reg),
-                      cpu.reg(src_reg), true);
-  } else {
-    // clang-format off
-    static constexpr std::array opcode_table = {
-        Opcode::And,
-        Opcode::Eor,
-        Opcode::Mov, // LSL
-        Opcode::Mov, // LSR
-        Opcode::Mov, // ASR
-        Opcode::Adc,
-        Opcode::Sbc,
-        Opcode::Mov, // ROR
-        Opcode::Tst,
-        Opcode::Rsb,
-        Opcode::Cmp,
-        Opcode::Cmn,
-        Opcode::Orr,
-        Opcode::Mov, // MUL
-        Opcode::Bic,
-        Opcode::Mvn,
-    };
-    // clang-format on
-    constexpr auto translated_opcode = opcode_table[opcode];
-    static_assert(translated_opcode != Opcode::Mov);
-
-    common::data_processing<translated_opcode, true>(
-        cpu, dest_reg,
-        cpu.reg(translated_opcode == Opcode::Rsb ? src_reg : dest_reg),
-        translated_opcode == Opcode::Rsb ? 0 : cpu.reg(src_reg),
-        cpu.program_status().carry());
-  }
-
-  return 1;
-}
-
-template <int opcode, bool hi_dest, bool hi_src>
-u32 hi_register_operation(Cpu& cpu, u16 instruction) {
-  const auto src_reg =
-      static_cast<Register>(((instruction >> 3) & 0b111) + hi_src * 8);
-  if constexpr (opcode == 0b11) {
-    // BX
-    common::branch_and_exchange(cpu, src_reg);
-    static_cast<void>(instruction);
-    return 3;
-  } else {
-    const auto dest_reg =
-        static_cast<Register>((instruction & 0b111) + hi_dest * 8);
-    static constexpr std::array opcode_table = {Opcode::Add, Opcode::Cmp,
-                                                Opcode::Mov};
-    constexpr auto translated_opcode = opcode_table[opcode];
-
-    common::data_processing<translated_opcode,
-                            translated_opcode == Opcode::Cmp>(
-        cpu, dest_reg, cpu.reg(dest_reg), cpu.reg(src_reg),
-        cpu.program_status().carry());
-  }
-  return 2;
-}
-
-template <Register dest_reg>
-u32 pc_relative_load(Cpu& cpu, u16 instruction) {
-  const u32 offset = (instruction & 0xff) << 2;
-
-  const u32 addr = ((cpu.reg(Register::R15) & ~0b10) + offset);
-
-  cpu.set_reg(dest_reg, cpu.mmu()->at<u32>(addr));
-
-  return cpu.mmu()->wait_cycles(addr, load_store_cycles(dest_reg, true));
-}
-
-template <bool load,
-          bool register_offset,
-          u32 register_or_offset,
-          typename Type>
-u32 load_store(Cpu& cpu, u16 instruction) {
-  const u32 offset = [&cpu]() -> u32 {
-    if constexpr (register_offset) {
-      return cpu.reg(static_cast<Register>(register_or_offset));
-    }
-    static_cast<void>(cpu);
-    return register_or_offset;
-  }();
-
-  const auto base_reg = static_cast<Register>((instruction >> 3) & 0b111);
-  const auto dest_reg = static_cast<Register>(instruction & 0b111);
-
-  const u32 addr = cpu.reg(base_reg) + offset;
-
-  constexpr auto mask = sizeof(Type) - 1;
-
-  const u32 aligned_addr = addr & ~mask;
-  const u32 resolved_addr = aligned_addr;
-  if constexpr (load) {
-    run_load<Type>(cpu, addr, dest_reg);
-  } else {
-    cpu.mmu()->set(resolved_addr, static_cast<Type>(cpu.reg(dest_reg)));
-  }
-  return cpu.mmu()->wait_cycles(resolved_addr,
-                                load_store_cycles(dest_reg, load));
-}
-
-template <bool load, Register dest_reg>
-u32 sp_relative_load_store(Cpu& cpu, u16 instruction) {
-  const u32 addr = cpu.reg(Register::R13) + ((instruction & 0xff) << 2);
-
-  if constexpr (load) {
-    const auto value = cpu.mmu()->at<u32>(addr & ~0b11);
-    cpu.set_reg(dest_reg, rotate_right(value, (addr & 0b11) * 8));
-  } else {
-    cpu.mmu()->set(addr, cpu.reg(dest_reg));
-  }
-
-  return cpu.mmu()->wait_cycles(addr, load_store_cycles(dest_reg, load));
-}
-
-template <bool sp, Register dest_reg>
-u32 load_address(Cpu& cpu, u16 instruction) {
-  constexpr auto base_reg = sp ? Register::R13 : Register::R15;
-  const u32 offset = ((instruction & 0xff) << 2);
-
-  const u32 value =
-      cpu.reg(base_reg) & (base_reg == Register::R15 ? ~0b10 : ~0);
-
-  cpu.set_reg(dest_reg, value + offset);
-
-  return 1;
-}
-
-template <bool negative>
-u32 add_offset_to_stack_pointer(Cpu& cpu, u16 instruction) {
-  const u32 offset = ((instruction & 0b111111) << 2);
-
-  const u32 sp_value = cpu.reg(Register::R13);
-  if constexpr (negative) {
-    cpu.set_reg(Register::R13, sp_value - offset);
-  } else {
-    cpu.set_reg(Register::R13, sp_value + offset);
-  }
-  return 1;
-}
-
-template <bool load, bool load_pc_or_store_lr>
-u32 push_pop_registers(Cpu& cpu, u16 instruction) {
-  u32 sp = cpu.reg(Register::R13);
-  u32 cycles = 0;
-
-  const auto add_cycles = [&cycles, &cpu](u32 addr, Register reg) {
-    cycles += cpu.mmu()->wait_cycles(addr, load_store_cycles(reg, load));
-  };
-
-  if constexpr (load) {
-    // POP/LDMIA
-    for (int i = 0; i < 8; ++i) {
-      if (test_bit(instruction, i)) {
-        const auto reg = static_cast<Register>(i);
-        add_cycles(sp, reg);
-        cpu.set_reg(reg, cpu.mmu()->at<u32>(sp));
-        sp += 4;
-      }
-    }
-    if constexpr (load_pc_or_store_lr) {
-      add_cycles(sp, Register::R15);
-      cpu.set_reg(Register::R15, cpu.mmu()->at<u32>(sp));
-      sp += 4;
-    }
-  } else {
-    // PUSH/STMDB
-    if constexpr (load_pc_or_store_lr) {
-      sp -= 4;
-      cpu.mmu()->set(sp, cpu.reg(Register::R14));
-      add_cycles(sp, Register::R14);
-    }
-    for (int i = 7; i >= 0; --i) {
-      if (test_bit(instruction, i)) {
-        sp -= 4;
-        const auto reg = static_cast<Register>(i);
-        cpu.mmu()->set(sp, cpu.reg(reg));
-        add_cycles(sp, reg);
-      }
-    }
-  }
-  cpu.set_reg(Register::R13, sp);
-  return cycles;
-}
-
-static int popcount(int num) {
-#ifdef __GNUC__
-  return __builtin_popcount(num);
-#else
-  int i = 0;
-  while (num) {
-    ++i;
-    num &= (num - 1);
-  }
-  return i;
-#endif
-}
-
-static int ctz(int num) {
-#ifdef __GNUC__
-  return __builtin_ctz(num);
-#else
-  int i = 0;
-  for (; i < 8 * sizeof(int); ++i) {
-    if (test_bit(num, i)) {
-      return i;
-    }
-  }
-  return i;
-#endif
-}
-
-template <bool load, Register base_reg>
-u32 multiple_load_store(Cpu& cpu, u16 instruction) {
-  u32 base = cpu.reg(base_reg);
-  u32 cycles = 1;
-
-  // Unspecified behavior for empty list
-  const auto reg_list = instruction & 0xff;
-  const bool empty_list = reg_list == 0;
-  if (load && empty_list) {
-    cpu.set_reg(Register::R15, cpu.mmu()->at<u32>(base));
-    base += 64;
-  } else if (!load && empty_list) {
-    cpu.mmu()->set(base, cpu.reg(Register::R15) + 2);
-    base += 64;
-  } else {
-    const auto count = popcount(reg_list);
-    const auto first_bit = ctz(reg_list);
-
-    for (int i = first_bit; i < 8; ++i) {
-      if (test_bit(instruction, i)) {
-        const auto reg = static_cast<Register>(i);
-        if constexpr (load) {
-          // LDMIA
-          cpu.set_reg(reg, cpu.mmu()->at<u32>(base));
-        } else {
-          // STMIA
-          if (reg == base_reg && i != first_bit) {
-            cpu.mmu()->set(base, cpu.reg(base_reg) + 4 * count);
-          } else {
-            cpu.mmu()->set(base, cpu.reg(reg));
-          }
-        }
-        cycles += cpu.mmu()->wait_cycles(
-            base, load_store_cycles(load ? reg : base_reg, load));
-
-        base += 4;
-      }
-    }
-  }
-
-  cpu.set_reg(base_reg, base);
-
-  return cycles;
-}
-
-u32 conditional_branch(Cpu& cpu, u16 instruction) {
-  // The condition is handled externally
-  const u32 offset =
-      static_cast<u32>(static_cast<s32>(static_cast<s8>(instruction & 0xff)))
-      << 1;
-
-  cpu.set_reg(Register::R15, cpu.reg(Register::R15) + offset);
-  return 1;
-}
-
-u32 software_interrupt(Cpu& cpu, u16 instruction) {
-  return execute_software_interrupt(cpu, static_cast<u32>(instruction & 0xff));
-}
-
-template <bool is_signed>
-u32 unconditional_branch(Cpu& cpu, u16 instruction) {
-  const u32 offset = (instruction & 0b111'1111'1111) << 1;
-  constexpr u32 mask = /*test_bit(offset, 11) */ is_signed ? 0xfffff800 : 0;
-  cpu.set_reg(Register::R15, cpu.reg(Register::R15) + (offset | mask));
-
-  return 3;
-}
-
-template <bool part_two>
-u32 long_branch_with_link(Cpu& cpu, u16 instruction) {
-  const u32 offset = (instruction & 0x7ff);
-  const u32 pc = cpu.reg(Register::R15) - 4;
-  if constexpr (!part_two) {
-    const s32 signed_offset =
-        (offset << 12) | (test_bit(offset, 10) ? 0xff800000 : 0);
-    cpu.set_reg(Register::R14, pc + signed_offset);
-    // cpu.set_reg(Register::R15, pc + 2);
-  } else {
-    const u32 next_instruction = pc + 2;
-    cpu.set_reg(Register::R15, cpu.reg(Register::R14) + (offset << 1) + 4);
-    cpu.set_reg(Register::R14, next_instruction | 1);
-  }
-  return 1;
-}
-
-[[noreturn]] u32 invalid_instruction([[maybe_unused]] Cpu& cpu,
-                                     [[maybe_unused]] u16 instruction) {
-  throw std::runtime_error("invalid instruction");
-}
-
-}  // namespace thumb
-
-using ThumbInstFunc = u32 (*)(Cpu&, u16);
-
-static constexpr std::array<ThumbInstFunc, 1024> thumb_lookup_table = [] {
-  std::array<ThumbInstFunc, 1024> res{};
-
-  const auto decode_thumb_instruction = [&](auto i) {
-    constexpr u16 instruction = i << 6;
-    const auto prefix = (instruction >> 13) & 0b111;
-    switch (prefix) {
-      case 0:
-        if (((instruction >> 11) & 0b11) == 0b11) {
-          constexpr bool immediate_operand = test_bit(instruction, 10);
-          constexpr bool subtract = test_bit(instruction, 9);
-          constexpr u32 register_or_value = (instruction >> 6) & 0b111;
-          res[i] = thumb::add_subtract<immediate_operand, subtract,
-                                       register_or_value>;
-        } else {
-          res[i] =
-              thumb::move_shifted_register<static_cast<ShiftType>(
-                                               ((instruction >> 11) & 0b11)),
-                                           (instruction >> 6) & 0b11111>;
-        }
-        break;
-      case 1: {
-        if constexpr (instruction > 0) {
-          // Move/compare/add/subtract immediate
-          constexpr auto decode_opcode = [](u16 inst) -> Opcode {
-            switch ((inst >> 11) & 0b11) {
-              case 0:
-                return Opcode::Mov;
-              case 1:
-                return Opcode::Cmp;
-              case 2:
-                return Opcode::Add;
-              case 3:
-                return Opcode::Sub;
-              default:
-                throw "Not reachable";
-            }
-          };
-          constexpr auto opcode = decode_opcode(instruction);
-          const auto dest_reg =
-              static_cast<Register>((instruction >> 8) & 0b111);
-
-          res[i] = thumb::move_compare_add_subtract_immediate<opcode, dest_reg>;
-        }
-        break;
-      }
-      case 2:
-        switch ((instruction >> 10) & 0b111111) {
-          case 0b010000:
-            // ALU operation
-            res[i] = thumb::alu_operation<(instruction >> 6) & 0b1111>;
-            break;
-          case 0b010001:
-            // Hi register operation
-            res[i] = thumb::hi_register_operation<((instruction >> 8) & 0b11),
-                                                  test_bit(instruction, 7),
-                                                  test_bit(instruction, 6)>;
-            break;
-          case 0b010010:
-          case 0b010011: {
-            //  Load/store
-            constexpr auto dest_reg =
-                static_cast<Register>((instruction >> 8) & 0b111);
-            res[i] = thumb::pc_relative_load<dest_reg>;
-            break;
-          }
-          default: {
-            constexpr bool sign_extended = test_bit(instruction, 9);
-            constexpr u32 offset_reg = (instruction >> 6) & 0b111;
-            if constexpr (sign_extended) {
-              constexpr auto opcode = (((instruction >> 10) & 0b1) << 1) |
-                                      ((instruction >> 11) & 0b1);
-
-              using Type = std::conditional_t<
-                  opcode <= 0b01, u16,
-                  std::conditional_t<opcode == 0b10, s8, s16>>;
-
-              constexpr bool load = opcode > 0;
-              res[i] = thumb::load_store<load, true, offset_reg, Type>;
-            } else {
-              constexpr bool load = test_bit(instruction, 11);
-              using Type =
-                  std::conditional_t<test_bit(instruction, 10), u8, u32>;
-              res[i] = thumb::load_store<load, true, offset_reg, Type>;
-            }
-            break;
-          }
-        }
-        break;
-      case 3: {
-        constexpr bool load = test_bit(instruction, 11);
-        constexpr bool byte_transfer = test_bit(instruction, 12);
-        constexpr u32 offset = ((instruction >> 6) & 0b11111)
-                               << (byte_transfer ? 0 : 2);
-        using Type = std::conditional_t<byte_transfer, u8, u32>;
-        res[i] = thumb::load_store<load, false, offset, Type>;
-
-        break;
-      }
-
-      case 4: {
-        constexpr bool is_sp_relative_store = test_bit(instruction, 12);
-        if constexpr (is_sp_relative_store) {
-          res[i] = nullptr;
-          constexpr bool load = test_bit(instruction, 11);
-          constexpr auto dest_reg =
-              static_cast<Register>((instruction >> 8) & 0b111);
-          res[i] = thumb::sp_relative_load_store<load, dest_reg>;
-        } else {
-          constexpr u32 offset = ((instruction >> 6) & 0b11111) << 1;
-          constexpr bool load = test_bit(instruction, 11);
-          res[i] = thumb::load_store<load, false, offset, u16>;
-        }
-        break;
-      }
-      case 5: {
-        if constexpr (!test_bit(instruction, 12)) {
-          constexpr bool sp = test_bit(instruction, 11);
-          constexpr auto dest_reg =
-              static_cast<Register>((instruction >> 8) & 0b111);
-          res[i] = thumb::load_address<sp, dest_reg>;
-        } else if constexpr (!test_bit(instruction, 10)) {
-          // Add offset to stack pointer
-
-          constexpr bool negative = test_bit(instruction, 7);
-          res[i] = thumb::add_offset_to_stack_pointer<negative>;
-        } else {
-          // Push pop
-          constexpr bool load = test_bit(instruction, 11);
-          constexpr bool load_pc_or_store_lr = test_bit(instruction, 8);
-          res[i] = thumb::push_pop_registers<load, load_pc_or_store_lr>;
-        }
-        break;
-      }
-      case 6: {
-        if constexpr (!test_bit(instruction, 12)) {
-          constexpr bool load = test_bit(instruction, 11);
-          constexpr auto base_reg =
-              static_cast<Register>((instruction >> 8) & 0b111);
-          res[i] = thumb::multiple_load_store<load, base_reg>;
-
-        } else {
-          if constexpr (((instruction >> 8) & 0b1111) == 0b1111) {
-            // Swi
-            res[i] = thumb::software_interrupt;
-          } else {
-            // Conditional branch
-            res[i] = thumb::conditional_branch;
-          }
-        }
-        break;
-      }
-      case 7: {
-        if (test_bit(instruction, 12)) {
-          constexpr bool part_two = test_bit(instruction, 11);
-          res[i] = thumb::long_branch_with_link<part_two>;
-        } else {
-          res[i] = thumb::unconditional_branch<test_bit(instruction, 10)>;
-        }
-        break;
-      }
-      default:
-        res[i] = thumb::invalid_instruction;
-        break;
-    }
-  };
-  for_static<2>([&](auto base_index) {
-    for_static<512>([&](auto i) {
-      constexpr auto real_index = (512 * decltype(base_index)::value) + i;
-      decode_thumb_instruction(
-          std::integral_constant<std::size_t, real_index>{});
-    });
-  });
-  return res;
-}();
-
+// using ThumbInstFunc = u32 (*)(Cpu&, u16);
 static constexpr std::array<InstFunc, 4096> arm_lookup_table = [] {
   using namespace arm;
   std::array<InstFunc, 4096> res{};
@@ -1968,8 +907,8 @@ static constexpr std::array<InstFunc, 4096> arm_lookup_table = [] {
 }();
 
 u32 Cpu::execute() {
-#if 0
-  if (interrupts_waiting.data() > 0 || halted) {
+#if 1
+  if (/*interrupts_waiting.data() > 0 ||*/ halted) {
     return 1;
   }
 #endif
@@ -1981,8 +920,10 @@ u32 Cpu::execute() {
     m_current_memory = storage;
     m_memory_offset = pc_region;
   }
+
   if (m_current_program_status.thumb_mode()) {
     const u32 pc = (m_regs[15] & ~0b1);
+    // fmt::printf("%08x\n", pc);
 
     u16 instruction;
     std::memcpy(&instruction, &m_current_memory[pc - m_memory_offset],
@@ -1997,11 +938,13 @@ u32 Cpu::execute() {
       }
       return 0;
     }
-    const auto inst_func = thumb_lookup_table[(instruction >> 6) & 0x3ff];
+    const auto inst_func =
+        thumb_instruction_tables.interpreter_table[(instruction >> 6) & 0x3ff];
     return inst_func(*this, instruction);
   }
 
   const u32 pc = (m_regs[15] & ~0b11);
+  // fmt::printf("%08x\n", pc);
   u32 arm_instruction;
   std::memcpy(&arm_instruction, &m_current_memory[pc - m_memory_offset],
               sizeof(u32));
